# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend Test Case API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/transaction/store": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Create Transaction",
                "operationId": "75a119bd9d0023e90ac8feb22bf59575",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "book_code": {
                                                "type": "string"
                                            },
                                            "member_code": {
                                                "type": "string"
                                            },
                                            "borrowed_at": {
                                                "type": "string"
                                            },
                                            "returned_at": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "book_code": "AB-12",
                                    "member_code": "M000",
                                    "borrowed_at": "2021-12-11",
                                    "returned_at": "2021-12-15"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "book_code": {
                                            "type": "string",
                                            "example": "AB-12"
                                        },
                                        "member_code": {
                                            "type": "string",
                                            "example": "M000"
                                        },
                                        "borrowed_at": {
                                            "type": "string",
                                            "example": "2021-12-11"
                                        },
                                        "returned_at": {
                                            "type": "string",
                                            "example": "2021-12-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaction/update/{transaction_id}": {
            "put": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Update Transaction for return the book has borrowed",
                "operationId": "8c88a825db408e19ccecf67263a61ec8",
                "parameters": [
                    {
                        "name": "transaction_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "transaction_id": {
                                                "type": "number"
                                            },
                                            "book_code": {
                                                "type": "string"
                                            },
                                            "member_code": {
                                                "type": "string"
                                            },
                                            "borrowed_at": {
                                                "type": "string"
                                            },
                                            "returned_at": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "transaction_id": "1",
                                    "book_code": "AB-12",
                                    "member_code": "M000",
                                    "borrowed_at": "2021-12-11",
                                    "returned_at": "2021-12-15"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "book_code": {
                                            "type": "string",
                                            "example": "AB-12"
                                        },
                                        "member_code": {
                                            "type": "string",
                                            "example": "M000"
                                        },
                                        "borrowed_at": {
                                            "type": "string",
                                            "example": "2021-12-11"
                                        },
                                        "returned_at": {
                                            "type": "string",
                                            "example": "2021-12-11"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}